Update lectureTopicTask quota to be course scoped instead of milestone scoped
Following a discussion of course grading with the professor, it was clarified lecture topic tasks are only tallied up at the end of the term. Because of this, the lectureTopicTaskQuota field will be kept in but no longer refer to a per milestone quota. The field will be reused at a later date to generate the end of course metrics.